lab7/requirements.txt
pyspark==3.5.1
findspark==2.0.1
pyyaml==6.0.1
pandas==2.2.2
matplotlib==3.8.4